+++ test/widgets/panel_interaction_extended_test.dart
@@ -26,7 +26,7 @@ void main() {
                   ),
                   TestContentPanel(
                     id: const PanelId('right'),
-                    flexOverride: 1,
+                    layoutWeightOverride: 1,
                     child: Container(),
                   ),
                 ],