+++ test/widgets/panel_interaction_test.dart
@@ -34,7 +34,7 @@ void main() {
                 ),
                 TestContentPanel(
                   id: const PanelId('right'),
-                  flexOverride: 1,
+                  layoutWeightOverride: 1,
                   child: Container(),
                 ),
               ],